% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAncestor.R
\name{goAncestor}
\alias{goAncestor}
\title{Title}
\usage{
goAncestor(gene2GO)
}
\arguments{
\item{gene2GO}{A dataframe with two columns: GeneID GOID}
}
\value{
A named list of GO IDs
}
\description{
Title
}
\examples{
For future
}
