% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhplot-package.R
\docType{package}
\name{hhplot-package}
\alias{hhplot}
\alias{hhplot-package}
\title{hhplot: My own plot and themes}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
