% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cafe.R
\name{plot_cafe}
\alias{plot_cafe}
\title{Draw CAFE5 result}
\usage{
plot_cafe(cafe, tree_95 = NULL, label2species = NULL)
}
\arguments{
\item{cafe}{File path. The CAFE result directory}

\item{tree_95}{(optional) File path. The tree generated from the MCMCTree software with 95% HPD}

\item{label2species}{(optional)  File path. A table with two columns used to replace labels in the tree. Two columns: Label/Species name}
}
\value{
A CAFE result tree
}
\description{
Draw CAFE5 result
}
\examples{
For future
}
