% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kegg.R
\name{plot_kegg}
\alias{plot_kegg}
\title{plot_kegg}
\usage{
plot_kegg(KO_list)
}
\arguments{
\item{KO_list}{a character vector. KO numbers.}
}
\description{
Draw the KEGG summary plot for genome analysis. Recommend save as 18 x 16 cm.
}
\examples{
For future

}
