% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme_Publication}
\usage{
theme_Publication(base_size = 8)
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\description{
theme_Publication
}
\examples{
For future
}
